Refactor TwoFactorAuthenticationRequest namespace
